Move unit test to Dialog.a11y
